Cut everything down to just the ncurses functions.
That's really all this crate should be. Garfanzo is kind of... another
project, I guess? I think I'll try to move it into some kind of example
directory that has rusty_hud as a dependency. The REASON for doing this,
by the way, is to allow other projects to use these functions, and to
let me add features a bit easier.